Reorganized diff engines into a directory.
